add correct month
перенес все в таск колум добавил условие не перехода на другой месяц и не перехода на день который уже прошел
