+++ src/components/TasksColumnsList/TasksColumnsList.jsx
@@ -3,9 +3,25 @@ import { COLUMNS } from 'constants/columns.constans';
 
 import TasksColumn from 'components/ChoosedDay/TasksColumn/TasksColumn';
 import scss from './TasksColumnsList.module.scss';
+import { useDispatch, useSelector } from 'react-redux';
+import { getCurrentDate } from 'redux/calendar/selectors';
+import { useTasksByChoosedDay } from 'hooks/useTasksByChoosedDay';
+import { fetchTasks } from 'redux/tasks/tasksOperations';
+import { useEffect } from 'react';
 
 
-const TasksColumnsList = ({ dayTasks }) => {
+const TasksColumnsList = () => {
+    
+  const dispatch = useDispatch();
+ const currentDate = useSelector(getCurrentDate);
+   useEffect(() => {
+     dispatch(fetchTasks());
+   }, [currentDate, dispatch]);
+  
+  const { tasks } = useTasksByChoosedDay(currentDate) || [];
+
+  const dayTasks = tasks;
+  
   const tasksByColumns = COLUMNS.reduce((acc, column) => {
     return {
       ...acc,
