+++ src/components/сalendar/Calendar.jsx
@@ -1,7 +1,7 @@
 
 
-import { useEffect } from "react";
 import css from "./Calendar.module.scss";
+import { useEffect } from "react";
 import { CalendarHeaderDays } from "./calendarHeaderDays/CalendarHeaderDays";
 import { CalendarGrid } from "./сalendarGrid/CalendarGrid";
 import { CalendarHeader } from "./сalendarHeader/CalendarHeader";
@@ -13,9 +13,10 @@ const Calendar = () => {
   const currentDate = useSelector(getCurrentDate);
       const dispatch = useDispatch();
   useEffect(() => {
+
     dispatch(fetchTasks(currentDate))
   },[currentDate, dispatch])
-// 
+
     return (
     <div className={css.calendar_wrapper}>
         <CalendarHeader/>
